% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{datauni}
\alias{datauni}
\title{Normalize and adjust data}
\usage{
datauni(data)
}
\arguments{
\item{data}{Input data}
}
\value{
Normalized data
}
\description{
Normalize and adjust data
}
\keyword{internal}
