% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{deltaanglefunc}
\alias{deltaanglefunc}
\title{Calculate delta angle for different rho values}
\usage{
deltaanglefunc(rho, mu, sigma, T, N, Tij, alpha, start)
}
\arguments{
\item{rho}{Correlation coefficient}

\item{mu}{Mean vector}

\item{sigma}{Variance vector}

\item{T}{Depth}

\item{N}{Branching factor}

\item{Tij}{Time parameter}

\item{alpha}{Alpha parameter (1 - coverage)}

\item{start}{Starting point}
}
\value{
Delta angle matrix
}
\description{
Calculate delta angle for different rho values
}
\keyword{internal}
