% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_functions.R
\name{generate_tree_data}
\alias{generate_tree_data}
\title{Generate Sample Tree Data}
\usage{
generate_tree_data(n_samples = 100, mu = c(0, 0), sigma = c(1, 1), rho = 0.5)
}
\arguments{
\item{n_samples}{Number of samples to generate. Default is 100.}

\item{mu}{Mean vector. Default is c(0, 0).}

\item{sigma}{Variance vector. Default is c(1, 1).}

\item{rho}{Correlation coefficient. Default is 0.5.}
}
\value{
Matrix with generated data
}
\description{
Generate sample tree-structured data for testing and demonstration.
}
\examples{
sample_data <- generate_tree_data(n_samples = 200, rho = 0.7)
plot(sample_data)

}
