% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{largefunc}
\alias{largefunc}
\title{Large-scale simulation function}
\usage{
largefunc(rho1, rho2, mu, sigma, T, N, Tij, n2, alpha, start)
}
\arguments{
\item{rho1}{First correlation coefficient}

\item{rho2}{Second correlation coefficient}

\item{mu}{Mean vector}

\item{sigma}{Variance vector}

\item{T}{Depth}

\item{N}{Branching factor}

\item{Tij}{Time parameter}

\item{n2}{Number of repetitions}

\item{alpha}{Alpha parameter}

\item{start}{Starting point}
}
\value{
Comparison results
}
\description{
Large-scale simulation function
}
\keyword{internal}
