% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{musigma}
\alias{musigma}
\title{Calculate mean and covariance}
\usage{
musigma(data)
}
\arguments{
\item{data}{Input data}
}
\value{
Vector with mean, variances, and correlation
}
\description{
Calculate mean and covariance
}
\keyword{internal}
