% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{rho12func}
\alias{rho12func}
\title{Compare two correlation coefficients}
\usage{
rho12func(rho1, rho2, T, N, Tij, n1, n2, alpha, start)
}
\arguments{
\item{rho1}{First correlation coefficient}

\item{rho2}{Second correlation coefficient}

\item{T}{Depth}

\item{N}{Branching factor}

\item{Tij}{Time parameter}

\item{n1}{Outer repetitions}

\item{n2}{Inner repetitions}

\item{alpha}{Alpha parameter}

\item{start}{Starting point}
}
\value{
Comparison results
}
\description{
Compare two correlation coefficients
}
\keyword{internal}
