% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{sigmamatrix}
\alias{sigmamatrix}
\title{Generate covariance matrix}
\usage{
sigmamatrix(sigma, rho, i)
}
\arguments{
\item{sigma}{Variance vector}

\item{rho}{Correlation coefficient}

\item{i}{Level parameter}
}
\value{
Covariance matrix
}
\description{
Generate covariance matrix
}
\keyword{internal}
