% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{split}
\alias{split}
\title{Split data generation}
\usage{
split(mu, sigma, rho, i, N, Tij)
}
\arguments{
\item{mu}{Mean vector}

\item{sigma}{Variance vector}

\item{rho}{Correlation coefficient}

\item{i}{Level parameter}

\item{N}{Branching factor}

\item{Tij}{Time parameter}
}
\value{
Generated data
}
\description{
Split data generation
}
\keyword{internal}
