% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generation.R
\name{treedataresidual}
\alias{treedataresidual}
\title{Generate tree-structured residual data}
\usage{
treedataresidual(mu, sigma, rho, T, N, Tij)
}
\arguments{
\item{mu}{Mean vector}

\item{sigma}{Variance vector}

\item{rho}{Correlation coefficient}

\item{T}{Depth}

\item{N}{Branching factor}

\item{Tij}{Time parameter}
}
\value{
Tree-structured data
}
\description{
Generate tree-structured residual data
}
\keyword{internal}
